requirements.txt
numpy==1.16.4
pandas==0.24.2
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tqdm==4.32.2
